backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Microservice for communicate with React"

requires-python = ">=3.12"
dependencies = [
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
]
